Refactor to use proper casing for component file name
