+++ src/components/Layout.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { Header } from '@components/header/header';
-import { Footer } from '@components/footer/footer';
+import { Header } from '@components/header/Header';
+import { Footer } from '@components/footer/Footer';
 import { useSiteMetadata } from '@hooks/use-site-metadata';
 import { Contact } from '@components/contact/Contact';
