--- src/components/offerings/Offerings.tsx
@@ -1,24 +1,0 @@
-import React from 'react';
-
-export const Offerings = () => {
-    return (
-        <>
-            <div className="offerings-container">
-                <div>
-                    <h2>Offerings section</h2>
-                    <p>Post-ironic kombucha single-origin coffee, salvia VHS portland iPhone hoodie selfies. Umami
-                        affogato woke shaman cray readymade single-origin coffee venmo twee green juice swag chambray
-                        hexagon pok pok. Franzen small batch iceland selfies etsy ethical salvia kitsch gorpcore
-                        microdosing 8-bit la croix JOMO celiac. La croix seitan taxidermy kogi VHS letterpress. Migas
-                        chillwave master cleanse, tacos blackbird spyplane snackwave big mood normcore four loko
-                        activated charcoal everyday carry taxidermy narwhal vexillologist cray. Echo park paleo tbh
-                        chartreuse, butcher taiyaki tonx scenester master cleanse tattooed hot chicken. Post-ironic
-                        williamsburg truffaut, fixie coloring book portland pok pok bushwick prism dreamcatcher hella
-                        lomo.
-                    </p>
-                    <a href="#">Learn more...</a>
-                </div>
-            </div>
-        </>
-    );
-};
